### 2018-04-18 #682 @whedon
[PRE REVIEW]: sumo: Command-line tools for plotting and analysis of ab initio calculations
**Submitting author:** @utf (<a href="http://orcid.org/0000-0002-4486-3321">Alex Ganose</a>)
**Repository:** <a href="https://github.com/SMTG-UCL/sumo" target ="_blank">https://github.com/SMTG-UCL/sumo</a>
**Version:** v1.0.4
