### 2018-04-18 f136e39
FIX: Better optplot legend handling
Fixes #4
