### 2018-04-18 0812354
FIX: Better fix for optplot
