### 2018-04-18 32eec78
FIX: Improve bandstats effective mass labels
Fixes #3
