### 2018-04-23 #682 @arfon
@kyleniemeyer - would you be willing to edit this submission for JOSS?
