### 2018-04-23 3666cb4
FIX: Fix density option in kgen
