+++ sumo/cli/kgen.py
@@ -105,7 +105,8 @@ def kgen(filename='POSCAR', directory=None, make_folders=False, symprec=0.01,
     poscar = Poscar.from_file(filename)
     kpath, kpoints, labels = get_path_data(poscar.structure, mode=mode,
                                            symprec=symprec, kpt_list=kpt_list,
-                                           labels=labels)
+                                           labels=labels,
+                                           line_density=density)
 
     logging.info('\nk-point label indicies:')
     for i, label in enumerate(labels):
