### 2018-04-26 f454573
FIX: Default optional arguments now work for band+dos plots
Bug originally mentioned in #7.
