### 2018-04-26 9dec5fb
FIX: Fix spg option in kgen and phonon-bandplot
