### 2018-04-26 7e641cf
FIX: Fix gaussian broadening for DOS plotting
