### 2018-05-02 #682 @kyleniemeyer
@jarvist great! I was actually about to email you, so thanks for volunteering. Knowing the authors is fine, as long as you don't have a formal conflict of interest based on our policy: https://github.com/openjournals/joss/blob/master/COI.md
