### 2018-05-02 #717 @utf
@kyleniemeyer: I've just had a look through the conflict of interest page you linked in the previous issue. Jarvist and David Scanlon (@scanlond), the last author on this paper, were on a paper together back in 2016: https://aip.scitation.org/doi/abs/10.1063/1.4943973
