### 2018-05-02 #717 @jarvist
I have however co-authored a few times with @ajjackson back when were were in the same group. If @kyleniemeyer is happy, I'm happy! I certainly do know these people, but I don't think that's any problem in terms of writing a fair and open review of these codes.
