### 2018-05-02 #717 @kyleniemeyer
@jarvist I can see the partial overlap—thanks for being upfront about it. I agree that these connections should not be an issue.
