### 2018-05-08 514082e
ENH: Cache DOS colours as they are created for consistent plots
