### 2018-05-08 765841b
TST: Unit testing and docstring reference for orbital colour cache
