### 2018-05-08 5e985d7
ENH: dosplot axis label options
