### 2018-05-08 ef560e1
ENH: ylabel and dos-label options for bandplot
