+++ sumo/plotting/bs_plotter.py
@@ -45,8 +45,11 @@ class SBSPlotter(BSPlotter):
         BSPlotter.__init__(self, bs)
 
     def get_plot(self, zero_to_efermi=True, ymin=-6., ymax=6.,
-                 width=6., height=6., vbm_cbm_marker=False, dpi=400, plt=None,
-                 dos_plotter=None, dos_options=None, dos_aspect=3, fonts=None):
+                 width=6., height=6., vbm_cbm_marker=False,
+                 ylabel='Energy (eV)',
+                 dpi=400, plt=None,
+                 dos_plotter=None, dos_options=None, dos_label=None,
+                 dos_aspect=3, fonts=None):
         """Get a :obj:`matplotlib.pyplot` object of the band structure.
 
         If the system is spin polarised, blue lines are spin up, red lines are
@@ -63,14 +66,15 @@ class SBSPlotter(BSPlotter):
             height (:obj:`float`, optional): The height of the plot.
             vbm_cbm_marker (:obj:`bool`, optional): Plot markers to indicate
                 the VBM and CBM locations.
+            ylabel (:obj:`str`, optional): y-axis (i.e. energy) label/units
             dpi (:obj:`int`, optional): The dots-per-inch (pixel density) for
                 the image.
             plt (:obj:`matplotlib.pyplot`, optional): A
                 :obj:`matplotlib.pyplot` object to use for plotting.
-            dos_plotter (:obj:`~sumo.plotting.dos_plotter.VDOSPlotter`, \
+            dos_plotter (:obj:`~sumo.plotting.dos_plotter.SDOSPlotter`, \
                 optional): Plot the density of states alongside the band
                 structure. This should be a
-                :obj:`~sumo.plotting.dos_plotter.VDOSPlotter` object
+                :obj:`~sumo.plotting.dos_plotter.SDOSPlotter` object
                 initialised with the data to plot.
             dos_options (:obj:`dict`, optional): The options for density of
                 states plotting. This should be formatted as a :obj:`dict`
@@ -108,6 +112,7 @@ class SBSPlotter(BSPlotter):
                         Plot the density of states for each element on seperate
                         subplots. Defaults to ``False``.
 
+            dos_label (:obj:`str`, optional): DOS axis label/unitsa
             dos_aspect (:obj:`float`, optional): Aspect ratio for the band
                 structure and density of states subplot. For example,
                 ``dos_aspect = 3``, results in a ratio of 3:1, for the band
@@ -157,18 +162,21 @@ class SBSPlotter(BSPlotter):
                 e = eners[nd][str(Spin.down)][nb]
                 ax.plot(dists[nd], e, 'r--', linewidth=band_linewidth)
 
-        self._maketicks(ax)
+        self._maketicks(ax, ylabel=ylabel)
         self._makeplot(ax, plt.gcf(), data, zero_to_efermi=zero_to_efermi,
                        vbm_cbm_marker=vbm_cbm_marker, width=width,
                        height=height, ymin=ymin, ymax=ymax,
-                       dos_plotter=dos_plotter, dos_options=dos_options)
+                       dos_plotter=dos_plotter, dos_options=dos_options,
+                       dos_label=dos_label)
         return plt
 
     def get_projected_plot(self, selection, mode='rgb', interpolate_factor=4,
                            circle_size=150, projection_cutoff=0.001,
                            zero_to_efermi=True, ymin=-6., ymax=6., width=6.,
-                           height=6., vbm_cbm_marker=False, dpi=400, plt=None,
-                           dos_plotter=None, dos_options=None,
+                           height=6., vbm_cbm_marker=False,
+                           ylabel='Energy (eV)',
[…]
