### 2018-05-09 e44495f
Release version 1.0.8
