### 2018-05-15 487aa80
BUG: bandstats: fix reading Eg from spin-pol calcs
Instead of always reporting the bandgap of the first spin channel,
check both and report the minimum. This is still not quite correct;
should examine the VBM and CBM of both channels.
