### 2018-05-21 80e7f8d
BUG: Fix phonon bandplot symprec option
