+++ CHANGELOG.rst
@@ -14,7 +14,8 @@ Enhancements by Adam Jackson:
 
 Various bug fixes:
 
-- Fixed gaussian broadening of DOS
+- Fixed for ``--symprec`` in phonon-bandplot.
+- Fixed gaussian broadening of DOS.
 - Fixed ``--spg`` option in kgen and phonon-bandplot.
 - Fixed default arguments for band structure + dos plotting.
 - Added A centered orthorhombic lattice to ``BradCrackKpath``.
