### 2018-05-29 #717 @cfgoldsmith
Thanks, Kyle. As we discussed earlier, I am not going to install the software, but will provide feedback on their objectives within the broader computational chemistry community. I don't see where to put that info. Also, since I'm not installing the software, would you prefer that I check the Funtionality boxes or leave them unchecked?
