### 2018-05-29 #717 @jarvist
Successfully installed & am slowly working my way through its functionality. It's quite a broad package with multiple features, which really requires some 'live' electronic structure calculations to throw it against.
