### 2018-05-29 #717 @kyleniemeyer
@cfgoldsmith yes, that's fine. Aside from checking off items, you can give your feedback to @utf here directly. As for the functionality items, you can just check off if there is evidence of them.
