### 2018-06-07 c8e6705
FIX: orbital projected band structures with branches
A pass-by-reference issue was removing projection data for
branches beyond the first. Deepcopy seems to get this working
as intended.
