### 2018-06-08 726b2fc
Finishing inplementing DOS to sumo-phonon-bandplot
