### 2018-06-11 ad8b7ed
ENH: Read phonon total_dos.dat without need to convert to CSV
