### 2018-06-11 3987e18
Remove commented code on phonon plotter
