### 2018-06-11 5ce0418
Phonon band plot: use matplotlib style for colour default
