### 2018-06-11 7faca1a
Phonon dos: replace "dos_file" syntax with "dos" for consistency
