+++ sumo/plotting/phonon_bs_plotter.py
@@ -36,12 +36,12 @@ class SPhononBSPlotter(PhononBSPlotter):
         PhononBSPlotter.__init__(self, bs)
         self.imag_tol = imag_tol
 
-    def _plot_phonon_dos(self, dos_file, ax=None, color=None):
+    def _plot_phonon_dos(self, dos, ax=None, color=None):
         if ax is None:
             ax = plt.gca()
         if color is None:
             color = 'C0'
-        dat = np.genfromtxt(dos_file, comments='#')
+        dat = np.genfromtxt(dos, comments='#')
         y, x = dat[:, 0], dat[:, 1]
         ax.plot(x, y, '-', color=color)
         ax.fill_betweenx(y, x, 0, color=color, alpha=0.5)
@@ -50,7 +50,7 @@ class SPhononBSPlotter(PhononBSPlotter):
         ax.set_xlabel("DOS")
 
     def get_plot(self, ymin=None, ymax=None, width=6., height=6., dpi=400,
-                 plt=None, fonts=None, dos_file=None, dos_aspect=3,
+                 plt=None, fonts=None, dos=None, dos_aspect=3,
                  color=None):
         """Get a :obj:`matplotlib.pyplot` object of the phonon band structure.
 
@@ -66,7 +66,7 @@ class SPhononBSPlotter(PhononBSPlotter):
                 specified as a :obj:`list` of :obj:`str`.
             plt (:obj:`matplotlib.pyplot`, optional): A
                 :obj:`matplotlib.pyplot` object to use for plotting.
-            dos_file (str): Path to total dos .csv data
+            dos (str): Path to total dos .csv data
             dos_aspect (float): Width division for vertical DOS
             color (:obj:`str` or :obj:`tuple`, optional): Line/fill colour in
                 any matplotlib-accepted format
@@ -78,7 +78,7 @@ class SPhononBSPlotter(PhononBSPlotter):
         if color is None:
             color = 'C0' # Default to first colour in matplotlib series
 
-        if dos_file:
+        if dos is not None:
             plt = pretty_subplot(1, 2, width, height, sharex=False, sharey=True, dpi=dpi,
                                  plt=plt, fonts=fonts,
                                  gridspec_kw={'width_ratios': [dos_aspect, 1],
@@ -103,14 +103,14 @@ class SPhononBSPlotter(PhononBSPlotter):
 
         self._maketicks(ax)
         self._makeplot(ax, plt.gcf(), data, width=width, height=height,
-                       ymin=ymin, ymax=ymax, dos_file=dos_file, color=color)
+                       ymin=ymin, ymax=ymax, dos=dos, color=color)
         plt.tight_layout()
         plt.subplots_adjust(wspace=0)
 
         return plt
 
     def _makeplot(self, ax, fig, data, ymin=None, ymax=None, height=6,
-                  width=6, dos_file=None, color=None):
+                  width=6, dos=None, color=None):
         """Utility method to tidy phonon band structure diagrams. """
 
         # Define colours
@@ -131,8 +131,8 @@ class SPhononBSPlotter(PhononBSPlotter):
         ax.set_xlim(0, data['distances'][-1][-1])
         ax.axhline(0, color=grey, linestyle='--')
 
-        if dos_file is not None:
-            self._plot_phonon_dos(dos_file, ax=fig.axes[1], color=color)
+        if dos is not None:
+            self._plot_phonon_dos(dos, ax=fig.axes[1], color=color)
         else:
 
             # keep correct aspect ratio square
