### 2018-06-12 f26f1a6
Phonon dos: read file within CLI logic, generate TDOS if no .dat given
