+++ sumo/cli/phonon_bandplot.py
@@ -18,6 +18,7 @@ TODO:
 """
 
 import os
+from os.path import isfile
 import sys
 import logging
 import argparse
@@ -197,6 +198,18 @@ def phonon_bandplot(filename, poscar=None, prefix=None, directory=None,
     bs = get_ph_bs_symm_line(yaml_file, has_nac=False,
                              labels_dict=kpath.kpoints)
 
+    # Replace dos filename with data array
+    if dos is not None:
+        if isfile(dos):
+            dos = np.genfromtxt(dos, comments='#')
+        elif dos:
+            phonon.set_mesh(qmesh, is_gamma_center=False, is_eigenvectors=True,
+                            is_mesh_symmetry=False)
+            phonon.set_total_DOS()
+            dos_freq, dos_val = phonon.get_total_DOS()
+            dos = np.zeros((len(dos_freq), 2))
+            dos[:, 0], dos[:, 1] = dos_freq, dos_val
+
     plotter = SPhononBSPlotter(bs)
     plt = plotter.get_plot(ymin=ymin, ymax=ymax, height=height, width=width,
                            plt=plt, fonts=fonts, dos=dos)
@@ -261,6 +274,7 @@ def _get_parser():
     parser.add_argument('-d', '--directory', metavar='D',
                         help='output directory for files')
     parser.add_argument('-q', '--qmesh', nargs=3, metavar='N',
+                        default=(8, 8, 8),
                         help='q-mesh to use for phonon DOS')
     parser.add_argument('-b', '--born', metavar='B',
                         help='born effective charge file')
@@ -308,7 +322,8 @@ def _get_parser():
     parser.add_argument('--dpi', type=int, default=400,
                         help='pixel density for image file')
     parser.add_argument('--font', default=None, help='font to use')
-    parser.add_argument('--dos', type=str, dest='dos', default=None,
+    parser.add_argument('--dos', nargs='?', type=str,
+                        default=None, const=True,
                         help='Phonopy .dat file for phonon DOS')
     return parser
 
