### 2018-06-19 79e2678
ENH: Default green for phonon bands/DOS, avoid clash with elec bands
