### 2018-06-19 8e75641
Update version in setup.py
