### 2018-06-20 6c19975
ENH: Rework DOS colour selection to use matplotlib style
