+++ sumo/plotting/dos_plotter.py
@@ -7,6 +7,7 @@ This module provides a class for plotting density of states data.
 """
 
 import itertools
+import matplotlib.pyplot
 
 from sumo.electronic_structure.dos import sort_orbitals
 from sumo.plotting import pretty_plot, pretty_subplot, colour_cycle
@@ -49,7 +50,7 @@ class SDOSPlotter(object):
                 }
 
             Usually generated using the
-            :obj:`sumo.electronric_structure.dos.get_pdos()` function.
+            :obj:`sumo.electronic_structure.dos.get_pdos()` function.
     """
 
     def __init__(self, dos, pdos=None):
@@ -57,7 +58,8 @@ class SDOSPlotter(object):
         self._pdos = pdos
 
     def dos_plot_data(self, yscale=1, xmin=-6., xmax=6., colours=None,
-                      plot_total=True, legend_cutoff=3, subplot=False):
+                      plot_total=True, legend_cutoff=3, subplot=False,
+                      cache=colour_cache):
         """Get the plotting data.
 
         Args:
@@ -86,6 +88,11 @@ class SDOSPlotter(object):
                 little contribution in the plotting range.
             subplot (:obj:`bool`, optional): Plot the density of states for
                 each element on seperate subplots. Defaults to ``False``.
+            cache (:obj:`dict`, optional): Cache object tracking how colours
+                have been assigned to orbitals. The format is the same as the
+                "colours" dict. This defaults to the module-level
+                sumo.plotting.colour_cache object, but an empty dict can be
+                used as a fresh cache. This object will be modified in-place.
 
         Returns:
             dict: The plotting data. Formatted with the following keys:
@@ -151,7 +158,8 @@ class SDOSPlotter(object):
                             for d in el_pdos[orb].densities.values()])
                 ymax = dmax if dmax > ymax else ymax
                 label = None if dmax < cutoff else '{} ({})'.format(el, orb)
-                colour = get_colour_for_element_and_orbital(el, orb, colours)
+                colour, cache = get_cached_colour(el, orb, colours,
+                                                  cache=cache)
                 el_lines.append({'label': label, 'alpha': 0.25,
                                  'colour': colour,
                                  'dens': el_pdos[orb].densities})
@@ -283,21 +291,19 @@ class SDOSPlotter(object):
 
         return plt
 
-
-def get_colour_for_element_and_orbital(element, orbital, colours=None):
+def get_cached_colour(element, orbital, colours=None, cache=colour_cache):
     """Get a colour for a particular elemental and orbital combination.
 
-    If the element is not specified in the colours dictionary, a random colour
-    will be used based on the list of 22 colours of maximum contast:
-    http://www.iscc.org/pdf/PC54_1724_001.pdf
+    If the element is not specified in the colours dictionary, the cache is
+    checked. If this element-orbital combination has not been chached before,
+    a new colour is drawn from the current matplotlib colour cycle and cached.
 
-    This is cached in sumo.plotting.colour_cache and re-used for subsequent
-    plots within this Python instance. To reset the cache, set
+    The default cache is sumo.plotting.colour_cache. To reset this cache, set
     ``sumo.plotting.colour_cache = {}``.
 
     Args:
-        element (str): The element.
-        orbital (str): The orbital.
+        element (:obj:`str`): The element.
+        orbital (:obj:`str`): The orbital.
         colours (:obj:`dict`, optional): Use custom colours for specific
             element and orbital combinations. Specified as a :obj:`dict` of
             :obj:`dict` of the colours. For example::
@@ -309,36 +315,56 @@ def get_colour_for_element_and_orbital(element, orbital, colours=None):
 
             The colour can be a hex code, series of rgb value, or any other
             format supported by matplotlib.
[…]
