### 2018-06-22 #717 @cfgoldsmith
Hi. Here is my review:
“sumo currently only supports VASP” this strikes me as rather curious, since VASP is one of the more expensive packages — even for academics. In the spirit of this journal, they should consider expanding their approach to include many of the open-source packages, such as GPAW, Orca, etc.
The authors fail to mention Molden, which blurs the line between command-line and gui, but is old and free and has much of the same functionality (though certainly not all).

The approach and examples is very specifically geared to solid state physics, even if they don’t explicitly say it. They should at least mention it. It would be helpful to elaborate on what future aspects they would consider, especially for the non solid-state community (e.g. what about electron density, molecular orbitals, charge distribution, etc).

Beyond that, the pack seems well developed, the graphics are appropriate and polished, and I think people will find it useful.
