### 2018-06-22 #717 @jarvist
Sumo does many different things - I haven't had time to try them all out yet! Here's a partial review, with some feedback.

1) It wasn't immediately obvious what the code was for + what it did. 'Ab-initio' is a very general term. Making it clear that these tools are for reciprocal-space band structures / computational solid-state physics would help.
Having used a few of the components, I think perhaps explicitly describing it as a convenience interface to the many dependent python packages / post-processing tools for VASP calculations is most descriptive. Sign-posting which of the many packages are being used 'under the hood' would also be useful,
i.e. something like
`An extensive framework for generating high-symmetry k-point paths,` by interfacing SeeK-path and SPGlib.
I also think the tutorials in the online documentation should be signposted more clearly in the README, as the /tests /examples folder don't offer much by themselves.
I noticed several spelling errors / typos in the STDOUT of the tools.

2) The disk space requirements are very heavy. It became apparent upon installing where the name 'sumo' came from, but it might be an idea to alert the user in the README.md.
I required the following additional packages to be able to install under pip3 on Debian:-
` sudo apt-get install python3-pip python3-numpy `
The lack of python3-numpy was a little cryptic, pip3 just failed at some point. Perhaps it can be added as a check?
The Debian additions required Required 72 MB (this was on top of the Python3 runtime).
Installed by pip3, I ended up with 361 MB in .local/lib/python3.5/site-packages/

3) Tests: I'm not really sure what to do with the tests directory.
Could you provide a 'runtests.py' in the root? Ideally you would add continuous integration. Reading the source of the tests, it's not obvious that they are tests, but more that they are examples where you can check by eye that some sensible results are produced.
As Sumo does little computation by itself, maybe this is acceptable, but it should be documented and explained. But you could e.g. have a computational pipeline + at least automatically validate that an output file is produced.

Suggested (partial review) actions:-
- [ ] Improve `README.md`
- [ ] Spell check sources, particularly the output to the user (for examples, try grep'ing for 'eleemtns' 'possibily' 'indicies:')
- [ ] Clean up / document / implement tests
