### 2018-06-22 #717 @utf
@cfgoldsmith

Thank you for your review.

We appreciate that VASP is a little exclusive which is unfortunate, however, it is very widely used in the field (there have been 25,000 papers published using the code in the last 10 years). Crucially VASP comes with a minimal toolchain, which motivated the development of this package.
We have intentionally designed Sumo to separate VASP-specific details from other elements in order to facilitate interfaces for open-source codes. Our team does not have much hands-on experience with the other codes that would most benefit from this, but contributions would be welcome from e.g. Abinit users to add these interfaces. We will add a section to the README which expressly invites this kind of assistance.

From what I can tell, Molden cannot directly load VASP output files and instead requires additional codes to convert the VASP output into a different format. Accordingly, I think Sumo and Molden occupy different niches. In particular, Molden is primarily focused on 3D visualisation which is not something we plan to support.

We agree that Sumo is currently geared towards the solid-state chemistry and physics communities (although the tools will work equally well for any molecular calculations performed in VASP). Sumo will likely remain focussed on the solid-state for the near future so we can make the mention of "solid-state" more prominent in the package descriptions if desired.

We appreciate that you found the package well developed and think others will find it useful. On this basis, are you able to complete the Review Checklist at the top of this page?
