### 2018-06-22 #717 @utf
@jarvist

Thank you for looking at Sumo.

In response to your numbered points:

1. We agree that ab initio is a general term. As noted in our response to @cfgoldsmith, we will update the documentation to make it clear that this package is geared for solid state analysis.
  - We already mention all the packages used under the hood in the README but perhaps we could be more specific where each of these is implemented.
  - Do you have a specific suggestion as to how we should signpost the tutorials more clearly? The README contains quite a few links to the documentation already.
  - Thanks for noticing the typos, we will go make sure to fix those.

2. For the disk space requirements, unfortunately there isn’t much we can do about that as most of this comes from the package dependencies. We will instead add a warning to the “Requirements” section of the readme to highlight this.

3. Regarding tests: we already have a section in the documentation detailing how to run the tests (see https://github.com/SMTG-UCL/sumo#tests). To be clear, these are definitely testing the code base and are not provided as examples of how to run the code.  Furthermore, we already have continuous integration enabled using Travis CI (this is indicated by the “build: passing” button at the top of the README). It isn't really expected that general users will poke around the test and example directories, as they are encouraged to install the package from PyPI.

Thank you again for reviewing Sumo.
