### 2018-06-22 ffc4792
FIX: Fix many typos
