### 2018-06-22 #717 @jarvist
To be clear - I'm certainly not requesting any extra development in response to my review, just more + clearer signposting in the documentation.

Apologies for missing the CI - would it be possible to add a code-coverage badge? I will investigate the unit tests next week. A two line README.md in the tests directory would not go amiss, repeating the info in the main README.md.
