### 2018-06-23 6d2375a
ENH: Add style file, fix dos tests
