+++ tests/tests_plotting/test_dos_plotter.py
@@ -8,16 +8,14 @@ try:
 except ImportError:
     import ConfigParser as configparser
 
-from sumo.plotting import colour_cycle, default_colours
+import matplotlib.pyplot
+
 import sumo.plotting.dos_plotter
-from sumo.plotting.dos_plotter import get_colour_for_element_and_orbital
+from sumo.plotting.dos_plotter import get_cached_colour
 
 
 class GetColourTestCase(unittest.TestCase):
     def setUp(self):
-        # Reboot the colour cycle for consistency
-        sumo.plotting.dos_plotter.col_cycle = colour_cycle()
-
         # Open default CLI colours config
         config_path = resource_filename(Requirement.parse('sumo'),
                                         'sumo/plotting/orbital_colours.conf')
@@ -27,43 +25,56 @@ class GetColourTestCase(unittest.TestCase):
 
     def test_get_colour_cache(self):
         """Check colour caching"""
-        col1 = tuple(get_colour_for_element_and_orbital('Hf', 's'))
-        col2 = tuple(get_colour_for_element_and_orbital('Zr', 'd'))
-        col3 = tuple(get_colour_for_element_and_orbital('Hf', 's'))
-
+        col1, cache = tuple(get_cached_colour('Hf', 's', cache=cache))
+        col2, cache = tuple(get_cached_colour('Zr', 'd', cache=cache))
+        col3, cache = tuple(get_cached_colour('Hf', 's', cache=cache))
         self.assertEqual(col1, col3)
         self.assertNotEqual(col1, col2)
+
+        # Try rebooting with new cache
+        col4, cache = tuple(get_cached_colour('Zr', 'd', cache={}))
+        self.assertEqual(col1, col4)
+
+    def test_get_colour_global_cache(self):
+        """Check colour caching"""
+        col1, cache = tuple(get_cached_colour('Hf', 's'))
+        col2, cache = tuple(get_cached_colour('Zr', 'd'))
+        col3, cache = tuple(get_cached_colour('Hf', 's'))
+        self.assertEqual(col1, col3)
+        self.assertNotEqual(col1, col2)
+
+        # Try rebooting with new cache
+        colour_cache={}
+        col4, cache = tuple(get_cached_colour('Zr', 'd'))
+        self.assertEqual(col1, col4)
 
     def test_get_colour_type_error(self):
         """Check bogus colour info is rejected"""
         with self.assertRaises(TypeError):
-            get_colour_for_element_and_orbital('Na', 'p', colours=('#aabbcc'))
+            get_cached_colour('Na', 'p', colours=('#aabbcc'))
 
     def test_get_colour_config(self):
         """Check orbital colours from config file"""
-        col_O_p = get_colour_for_element_and_orbital('O', 'p',
-                                                     colours=self.config)
-        col_Re_d = get_colour_for_element_and_orbital('Re', 'd',
-                                                      colours=self.config)
+        col_O_p, _ = get_cached_colour('O', 'p', colours=self.config)
+        col_Re_d, _ = get_cached_colour('Re', 'd', colours=self.config)
 
         self.assertEqual(col_O_p, '#0DB14B')
         self.assertEqual(col_Re_d, '#A154A1')
 
     def test_get_colour_mixed(self):
         """Check new colours drawn in correct sequence"""
-        col_O_p = get_colour_for_element_and_orbital('O', 'p',
-                                                     colours=self.config)
-        col_Hf_s = get_colour_for_element_and_orbital('Hf', 's',
-                                                      colours=self.config)
-        col_Re_d = get_colour_for_element_and_orbital('Re', 'd',
-                                                      colours=self.config)
[…]
