### 2018-06-24 #717 @kyleniemeyer
@jarvist @cfgoldsmith thanks for your reviews!

@utf please let us know when you've made changes in response to their comments, and specifically what changes you made. (If you open up issues or PRs on your own repository in response, you can link to them here to make it even easier.)
