### 2018-06-25 dc093ea
TST: Fixed DOS colour tests, updated cache clear instructions
