### 2018-06-25 bc4d980
ENH: dosplot using matplotlib styling. Other plots probably broken.
Two new parameters are made available at the command-line interface
for specifying additional style files and for optionally suppressing
the base style. To get a white-on-black plot for presentation, all
the user needs to supply is "--style dark_background".

Because styling is now performed at the "script"
level, plots through the Python API should follow the current MPL
context. This adds slightly more overhead for reproducing standard
sumo plots though the API, but should reduce the number of formatting
"surprises" when plo
[…]
