### 2018-06-25 47339bd
ENH: Total DOS line follow text colour if available
