### 2018-06-25 c65c038
Fix DOS subplot mode, use axes.labelcolor for ylabel
