### 2018-06-25 3d0e533
Wrap bandplot in style context
