### 2018-06-25 7c27f92
ENH: Bandplot using matplotlib styling
