### 2018-06-25 6981bcc
Clean up bandplot colour handling
