### 2018-06-25 5274180
Move optics plot colours to style sheet
