### 2018-06-25 ce0e796
ENH: Optics plotter using matplotlib styles
