+++ sumo/cli/phonon_bandplot.py
@@ -21,6 +21,7 @@ import os
 from os.path import isfile
 import sys
 import logging
+from pkg_resources import resource_filename
 import argparse
 import numpy as np
 
@@ -30,6 +31,7 @@ warnings.filterwarnings("ignore", category=FutureWarning,
 
 import matplotlib as mpl
 mpl.use('Agg')
+from matplotlib.pyplot import style as mpl_style
 
 from phonopy.units import VaspToTHz
 
@@ -53,6 +55,7 @@ def phonon_bandplot(filename, poscar=None, prefix=None, directory=None,
                     primitive_axis=None, line_density=60,
                     symprec=0.01, mode='bradcrack', kpt_list=None,
                     eigenvectors=False, labels=None, height=6., width=6.,
+                    style=None, no_base_style=False,
                     ymin=None, ymax=None, image_format='pdf', dpi=400,
                     plt=None, fonts=None, dos=None):
     """A script to plot phonon band structure diagrams.
@@ -120,6 +123,10 @@ def phonon_bandplot(filename, poscar=None, prefix=None, directory=None,
         width (:obj:`float`, optional): The width of the plot.
         ymin (:obj:`float`, optional): The minimum energy on the y-axis.
         ymax (:obj:`float`, optional): The maximum energy on the y-axis.
+        style (:obj:`list` or :obj:`str`, optional): (List of) matplotlib style
+            specifications, to be composed on top of Sumo base style.
+        no_base_style (:obj:`bool`, optional): Prevent use of sumo base style.
+            This can make alternative styles behave more predictably.
         image_format (:obj:`str`, optional): The image file format. Can be any
             format supported by matplotlib, including: png, jpg, pdf, and svg.
             Defaults to pdf.
@@ -210,23 +217,37 @@ def phonon_bandplot(filename, poscar=None, prefix=None, directory=None,
             dos = np.zeros((len(dos_freq), 2))
             dos[:, 0], dos[:, 1] = dos_freq, dos_val
 
-    plotter = SPhononBSPlotter(bs)
-    plt = plotter.get_plot(ymin=ymin, ymax=ymax, height=height, width=width,
-                           plt=plt, fonts=fonts, dos=dos)
-
-    if save_files:
-        basename = 'phonon_band.{}'.format(image_format)
-        filename = '{}_{}'.format(prefix, basename) if prefix else basename
-
-        if directory:
-            filename = os.path.join(directory, filename)
-
-        plt.savefig(filename, format=image_format, dpi=dpi,
-                    bbox_inches='tight')
-
-        filename = save_data_files(bs, prefix=prefix, directory=directory)
+    if style is None:
+        style = []
+    elif type(style) == str:
+        style = [style]
+    if no_base_style:
+        base_style = []
     else:
-        return plt
+        base_style = [resource_filename('sumo.plotting', 'sumo_base.mplstyle'),
+                      resource_filename('sumo.plotting', 'sumo_bs.mplstyle'),
+                      resource_filename('sumo.plotting', 'sumo_phonon.mplstyle'
+                                        )]
+
+    with mpl_style.context(base_style + style):
+
+        plotter = SPhononBSPlotter(bs)
+        plt = plotter.get_plot(ymin=ymin, ymax=ymax, height=height,
+                               width=width, plt=plt, fonts=fonts, dos=dos)
+
+        if save_files:
+            basename = 'phonon_band.{}'.format(image_format)
+            filename = '{}_{}'.format(prefix, basename) if prefix else basename
+
+            if directory:
+                filename = os.path.join(directory, filename)
+
+            plt.savefig(filename, format=image_format, dpi=dpi,
+                        bbox_inches='tight')
+
[…]
