### 2018-06-25 ca3bb51
Rework width/height in phonon plot to rely on style if not specified
