### 2018-06-25 2f8a073
Remove magic numbers from optics plots
