### 2018-06-25 de61b16
Remove magic numbers from bandplot
