### 2018-06-29 b126259
ENH: use scheme colours for spin channels; adjust tick handling
- Show y ticks on both sides of band structure
- hide minor ticks by default
- suppress x ticks on band structure and use thicker gridlines
- Fix ordering so yticks appear above other elements
