### 2018-06-29 3d4d94a
Don't override "grid" colours, do it with styles
