### 2018-07-02 2712d70
Tweak bandplot grid ordering; draw matching vertical lines on top
Band structures are special; the "grid" of x-values should separate
regions of the band structure, while the y-values should like behind
the data as a guide. To implement this while following style settings,
additional vertical lines are drawn following the grid style.

This applies to electronic and phonon band structures.

Unrelated change:
The dashed line for 0-frequency in phonon band structures is also
extended into the DOS region if it exists.
