### 2018-07-02 2022d41
BUG: Fix height/width handling in band structures
Aspect ratio now follows the canvas shape correctly
