### 2018-07-02 8d866fa
BUGFIX: Aspect ratio handling in optics plot
