+++ sumo/plotting/optics_plotter.py
@@ -150,7 +150,10 @@ class SOpticsPlotter(object):
 
         x0, x1 = ax.get_xlim()
         y0, y1 = ax.get_ylim()
-        height, width = rcParams['figure.figsize']
+        if width is None:
+            width = rcParams['figure.figsize'][0]
+        if height is None:
+            height = rcParams['figure.figsize'][1]
         ax.set_aspect((height/width) * ((x1-x0)/(y1-y0)))
         plt.tight_layout()
 
