### 2018-07-06 e79edef
DOC: Rework README in response to JOSS review comments
- Emphasise solid-state scope
- Emphasise use of other Python packages
- Add Contributing guidelines
- Raise links to Tutorial and Gallery
- Rework installation/requirements notes
- Add README to test directory
