+++ README.rst
@@ -9,25 +9,41 @@ Sumo
     :alt: Documentation Status
 
 Sumo is a Python toolkit for plotting and analysis of ab initio
-calculation data. The main features include:
+solid-state calculation data,
+built on existing Python packages from the solid-state
+chemistry/physics community.
+It is hoped that these command-line tools will bring some of the
+benefits of these libraries to a wider user-base while providing
+publication-ready plotting (powered by Matplotlib_.)
+
+The main features include:
 
 1. An extensive framework for generating high-symmetry k-point paths.
+
+   - Crystallographic spacegroups are determined using Spglib_.
+   - Conventional crystallographic paths are built in as well as interfaces to
+     the SeeK-path_ and
+     Pymatgen_ implementations.
+
 2. Plotting scripts for electronic and phonon band structures, density
    of states, and optical absorption diagrams.
+
+   - VASP calculations are imported using Pymatgen_.
+   - The Phonopy_ framework is supported for phonon band structures.
+
 3. Analysis scripts to calculate parabolic and non-parabolic band
    effective masses.
 
-The code currently only supports VASP calculations, however, we plan to
-add support for additional codes in future releases.
+   - Curve fitting is performed using `Scipy <https://www.scipy.org>`_.
+
+The code currently only supports VASP calculations, however, we would
+like to add support for additional solid-state codes in future
+releases. Code contributions to interface with these packages are
+welcome.
 
 Sumo is free to use, however, we ask that you cite the code if you use
-it in your research.
-
-Warning: There are probably still some bugs. If you think you've found
-one, please report it on the `Issue Tracker
-<https://github.com/SMTG-UCL/sumo/issues>`_. We welcome your help in
-improving and extending the package with your own contributions.
-
+it in your research. See the "contributing" section for information
+about reporting bugs and getting involved.
 
 Usage
 -----
@@ -39,7 +55,13 @@ tutorials and API documentation, is `available online
 help (``-h``) option for each command provides a summary of the
 available options.
 
-Currently, the scripts provided by sumo are:
+A guide to using each command can be found on the
+`Tutorial page <http://sumo.readthedocs.io/en/latest/tutorials.html>`_.
+
+For a preview of the functionality of sumo, see the
+`Gallery <http://sumo.readthedocs.io/en/latest/gallery.html>`_.
+
+Currently, the scripts provided are:
 
 - ``sumo-kgen``: For generating VASP KPOINTS files along high-symmetry
   k-point paths.
@@ -54,25 +76,29 @@ Currently, the scripts provided by sumo are:
 - ``sumo-bandstats``: For calculating electron and hole effective masses
   from a band structure.
 
-A guide to using each command can be found on the
-`Tutorial page <http://sumo.readthedocs.io/en/latest/tutorials.html>`_.
-
-For a preview of the functionality of sumo, see the
-`Gallery <http://sumo.readthedocs.io/en/latest/gallery.html>`_.
-
 
 Installation
 ------------
 
-We recommend installation from source with Pip, this will automatically
-install any dependencies:
+Sumo is a Python 3 package and requires a
+`typical scientific Python stack <https://www.scipy.org/about.html>`_;
+we recommend using your main package manager if possible
+(e.g. apt, Homebrew), or Anaconda to install Python 3 with setuptools.
+It is a good idea to also use this package manager to install
+Numpy and Matplotlib, as building them with setuptools can be troublesome.
+Sumo can then be installed using the Python package manager "Pip",
+which will automatically setup other Python packages as required:
 
 .. code-block:: bash
 
     pip3 install --user sumo
 
[…]
