### 2018-07-06 3c4ef00
Update README.rst
