### 2018-07-06 #717 @ajjackson
I have pushed an update to master that should address these review points. @utf has already assaulted the collection of typos.
Summary of changes pasted here from the commit message:
```
- Emphasise solid-state scope
- Emphasise use of other Python packages
- Add Contributing guidelines
- Raise links to Tutorial and Gallery
- Rework installation/requirements notes
- Add README to test directory
```
And "ab initio calculations" in the setup.py (i.e. for PyPI description) is replaced with "ab initio solid-state calculations".
