### 2018-07-06 #717 @jarvist
With these edits, `README.rst` is much more clear, and I think you have covered all of my partial review points.

Running your unit tests as your documented
`python3 -m unittest discover tests`
in the pip-installed `~/.local/lib/python3.5/site-packages/sumo/` (maybe point users in this direction in the README?), resulted in:
```
Ran 97 tests in 3.198s

FAILED (failures=4, errors=14)
```
(Full STDOUT here: https://gist.github.com/jarvist/ae461f6dd22e1f989a0625bbf4b63f3f )

Most of the errors looked like it trying to find data sets (bz2 / gz) files, which might be hard-coded / not present in the repository?
