### 2018-07-06 #717 @ajjackson
Running from inside `~/.local/lib` is rather unorthodox, does it work from a cloned source directory?
