### 2018-07-10 8ba37da
PKG: rename RTD requirements
