+++ readthedocs.yml
@@ -3,3 +3,5 @@ python:
     pip_install: true
     extra_requirements:
          - docs
+
+requirements_file: requirements_rtd.txt
