### 2018-07-10 2a12ddb
PKG: update references to requirements.txt
