### 2018-07-11 304b121
STY: Cleanup and improvements based on Codacy review
