### 2018-07-11 d635836
STY: More Codacy tweaks that slipped through
