### 2018-07-11 aed91f8
Remove some leftover cruft
