### 2018-07-11 e7d57b8
ENH: Use decorator to apply styling to dos plot
