### 2018-07-11 #717 @ajjackson
@jarvist are the tests working for you when they are run from the source folder? In the README we say "the unittests can be run (from the root directory of the project) ..."

The Travis logs show the process followed there, which simply clones the repo https://travis-ci.org/SMTG-UCL/sumo/jobs/402148229
