### 2018-07-12 #717 @jarvist
I'm travelling this week, so haven't had the opportunity to re-install a 'developer' version + try the tests from there.

It seems a bit strange that the tests ran partially. They seemed to be failing either because the datasets were not downloading(?) (permissions in the pip folder?), or that a python dependency on BZip2/Gzip was not present (assumed present but not specified in the package?).
