### 2018-07-12 #717 @ajjackson
Looking at the traceback more carefully, I don't think the tests you found are anything to do with Sumo. Sumo does not install tests to lib, and most of the messages refer to something called "monty"?
