### 2018-07-12 ff0903e
Tidy up styling implementation
