### 2018-07-12 39f1ed4
Use decorator for plot styling
