### 2018-07-13 01a2e88
BUG: Fixed bug setting font
