### 2018-07-13 7e0b5e1
Standardise plotting style
