### 2018-07-13 64dfaa5
STY: Fix codacy issues
