### 2018-07-16 9b06205
More codacy fixes
