### 2018-07-24 #717 @ajjackson
I believe we have now responded to all reviewer points, except for the suggestion to include a code-coverage badge. This is something we have considered, but it is going to be difficult to achieve _meaningful_ high coverage in a package that a) mostly consists of command line tools and b) mostly outputs graphics. At present the tests are focused on the internal data manipulation. It's achievable with some thought and effort, but not a priority for us at this point.
