### 2018-07-24 #717 @kyleniemeyer
Hi @jarvist I noticed a few of your checklist items unfilled above—are you satisfied with the changes they have made?
