### 2018-07-25 ca7be84
Move plotting font options to style decorator from pretty plots
