+++ sumo/plotting/phonon_bs_plotter.py
@@ -94,12 +94,11 @@ class SPhononBSPlotter(PhononBSPlotter):
         if dos is not None:
             plt = pretty_subplot(1, 2, width=width, height=height,
                                  sharex=False, sharey=True, dpi=dpi, plt=plt,
-                                 fonts=fonts,
                                  gridspec_kw={'width_ratios': [dos_aspect, 1],
                                               'wspace': 0})
             ax = plt.gcf().axes[0]
         else:
-            plt = pretty_plot(width, height, dpi=dpi, plt=plt, fonts=fonts)
+            plt = pretty_plot(width, height, dpi=dpi, plt=plt)
             ax = plt.gca()
 
         data = self.bs_plot_data()
