### 2018-07-27 bce2f9c
Changing instances of absorption to loss
