+++ sumo/plotting/optics_plotter.py
@@ -158,3 +158,133 @@ class SOpticsPlotter(object):
         plt.tight_layout()
 
         return plt
+
+class SLossPlotter(object):
+    """Class for plotting optical loss spectra.
+
+    The easiest way to initialise this class is using the output from the
+    :obj:`sumo.electronic_structure.optics.calculate_alpha()` method.
+
+    Args:
+        loss_data (:obj:`tuple` or :obj:`list`): The optical loss
+            spectra. Should be formatted as a :obj:`tuple` of :obj:`list`::
+
+                ([energies], [loss])
+
+            Alternatively, the anisotropic (directional dependent) loss
+            can be plotted if the data formatted as::
+
+                ([energies], [loss_xx, loss_yy, loss_zz])
+
+            If a :obj:`list` of :obj:`tuple` is provided, then multiple
+            loss spectra can be plotted simultaneously.
+        band_gap (:obj:`float` or :obj:`list`, optional): The band gap as a
+            :obj:`float`, plotted as a dashed line. If plotting multiple
+            spectra then a :obj:`list` of band gaps can be provided.
+        label (:obj:`str` or :obj:`list`): A label to identify the spectra.
+            If plotting multiple spectra then a :obj:`list` of labels can
+            be provided.
+    """
+
+    def __init__(self, loss_data, band_gap=None, label=None):
+        if type(loss_data) is tuple:
+            loss_data = [loss_data]
+
+        if type(band_gap) is float:
+            band_gap = [band_gap]
+        elif not band_gap:
+            band_gap = [None] * len(loss_data)
+
+        if type(label) is str:
+            label = [label]
+        elif not label and len(loss_data) > 1:
+            label = [str(i) for i in range(1, len(loss_data) + 1)]
+        elif not label:
+            label = [''] * len(loss_data)
+
+        if len(set([len(label), len(loss_data), len(band_gap)])) > 1:
+            raise ValueError('loss_data, band_gap, and label not same size')
+
+        self._loss_data = loss_data
+        self._band_gap = band_gap
+        self._label = label
+        self._xmax = max(loss_data[0][0])
+
+    def get_plot(self, width=6., height=6., xmin=0., xmax=None, ymin=0,
+                 ymax=2., colours=None, dpi=400, plt=None, fonts=None):
+        """Get a :obj:`matplotlib.pyplot` object of the optical spectra.
+
+        Args:
+            width (:obj:`float`, optional): The width of the plot.
+            height (:obj:`float`, optional): The height of the plot.
+            xmin (:obj:`float`, optional): The minimum energy on the x-axis.
+            xmax (:obj:`float`, optional): The maximum energy on the x-axis.
+            ymin (:obj:`float`, optional): The minimum loss intensity on
+                the y-axis.
+            ymax (:obj:`float`, optional): The maximum loss intensity on
+                the y-axis.
+            colours (:obj:`list`, optional): A :obj:`list` of colours to use in
+                the plot. The colours can be specified as a hex code, set of
+                rgb values, or any other format supported by matplotlib.
+            dpi (:obj:`int`, optional): The dots-per-inch (pixel density) for
+                the image.
+            plt (:obj:`matplotlib.pyplot`, optional): A
+                :obj:`matplotlib.pyplot` object to use for plotting.
+            fonts (:obj:`list`, optional): Fonts to use in the plot. Can be a
+                a single font, specified as a :obj:`str`, or several fonts,
+                specified as a :obj:`list` of :obj:`str`.
+
+        Returns:
+            :obj:`matplotlib.pyplot`: The plot of optical spectra.
+        """
[…]
