+++ sumo/cli/lossplot.py
@@ -1,0 +1,206 @@
+# coding: utf-8
+# Copyright (c) Scanlon Materials Theory Group
+# Distributed under the terms of the MIT License.
+
+"""
+A script to calculate and plot optical loss spectra from a VASP calculation.
+"""
+
+import os
+import sys
+import logging
+import warnings
+import argparse
+sys.path.insert(0, '/Users/keithbutler/src/sumo/')
+import matplotlib as mpl
+mpl.use('Agg')
+import sumo
+
+from pymatgen.io.vasp import Vasprun
+from pymatgen.util.string import latexify
+
+from sumo.plotting.optics_plotter import SLossPlotter
+from sumo.electronic_structure.optics import (broaden_eps, calculate_loss,
+                                              write_files)
+
+__author__ = "Keith Butler"
+__version__ = "1.0"
+__maintainer__ = "?"
+__email__ = ""
+__date__ = "July 25, 2018"
+
+
+def lossplot(filenames=None, prefix=None, directory=None,
+            gaussian=None, band_gaps=None, labels=None, average=True, height=6,
+            width=6, xmin=0, xmax=None, ymin=0, ymax=2.0, colours=None,
+            image_format='pdf', dpi=400, plt=None, fonts=None):
+    """A script to plot optical loss spectra from VASP calculations.
+
+    Args:
+        filenames (:obj:`str` or :obj:`list`, optional): Path to vasprun.xml
+            file (can be gzipped). Alternatively, a list of paths can be
+            provided, in which case the loss spectra for each will be
+            plotted concurrently.
+        prefix (:obj:`str`, optional): Prefix for file names.
+        directory (:obj:`str`, optional): The directory in which to save files.
+        gaussian (:obj:`float`): Standard deviation for gaussian broadening.
+        band_gaps (:obj:`float` or :obj:`list`, optional): The band gap as a
+            :obj:`float`, plotted as a dashed line. If plotting multiple
+            spectra then a :obj:`list` of band gaps can be provided.
+        labels (:obj:`str` or :obj:`list`): A label to identify the spectra.
+            If plotting multiple spectra then a :obj:`list` of labels can
+            be provided.
+        height (:obj:`float`, optional): The height of the plot.
+        width (:obj:`float`, optional): The width of the plot.
+        xmin (:obj:`float`, optional): The minimum energy on the x-axis.
+        xmax (:obj:`float`, optional): The maximum energy on the x-axis.
+        ymin (:obj:`float`, optional): The minimum loss intensity on the
+            y-axis.
+        ymax (:obj:`float`, optional): The maximum loss intensity on the
+            y-axis.
+        colours (:obj:`list`, optional): A :obj:`list` of colours to use in the
+            plot. The colours can be specified as a hex code, set of rgb
+            values, or any other format supported by matplotlib.
+        image_format (:obj:`str`, optional): The image file format. Can be any
+            format supported by matplotlib, including: png, jpg, pdf, and svg.
+            Defaults to pdf.
+        dpi (:obj:`int`, optional): The dots-per-inch (pixel density) for
+            the image.
+        plt (:obj:`matplotlib.pyplot`, optional): A
+            :obj:`matplotlib.pyplot` object to use for plotting.
+        fonts (:obj:`list`, optional): Fonts to use in the plot. Can be a
+            a single font, specified as a :obj:`str`, or several fonts,
+            specified as a :obj:`list` of :obj:`str`.
+
+    Returns:
+        A matplotlib pyplot object.
+    """
+    if not filenames:
+
+        if os.path.exists('vasprun.xml'):
+            filenames = ['vasprun.xml']
+        elif os.path.exists('vasprun.xml.gz'):
[…]
