### 2018-07-27 e18d0f0
Automatically determine ymax for loss plot
