### 2018-07-27 ff16c4b
Add cli installation for sumo-loss
