### 2018-07-27 8fba54a
automatically determine ymax for anisotropic plotting of loss function
