### 2018-07-27 7dd63e3
Fixed some codacy issues
