### 2018-07-30 0e55bc6
aligns DOS to fermi
