+++ sumo/plotting/dos_plotter.py
@@ -87,7 +87,7 @@ class SDOSPlotter(object):
                 little contribution in the plotting range.
             subplot (:obj:`bool`, optional): Plot the density of states for
                 each element on separate subplots. Defaults to ``False``.
-            zero_to_efermi (:obj:`bool`): Normalise the plot such that the
+            zero_to_efermi (:obj:`bool`, optional): Normalise the plot such that the
                 valence band maximum is set as 0 eV.
 
         Returns:
@@ -126,7 +126,9 @@ class SDOSPlotter(object):
         dos = self._dos
         pdos = self._pdos
         mask = (dos.energies >= xmin - 0.05) & (dos.energies <= xmax + 0.05)
-        plot_data = {'mask': mask, 'energies': dos.energies - dos.efermi if zero_to_efermi else dos.energies}
+        plot_data = {'mask': mask,
+                     'energies': dos.energies - dos.efermi
+                     if zero_to_efermi else dos.energies}
         spins = dos.densities.keys()
         ymax = 0
 
