### 2018-08-02 #717 @ajjackson
@jarvist did you have any more issues or can we wrap this up?
