### 2018-08-03 #717 @jarvist
If @ajjackson could add some instructions in the README / etc. for how a standard user (following the supplied installation instructions) can run the tests, then I'll consider it closed.

I'm generally happy with the package; I'll now go and review the `paper.md`.
