### 2018-08-03 #717 @utf
Hi @jarvist, standard users are not expected to run the tests. If you install from pip directly using the `pip3 install --user sumo` command the test data will not be installed (as is standard for all pip installations).

Developers are expected to run the tests through the instructions in the README.

To make this clear will we state that the developer installation should occur from a clone of the repository, and that tests can only be run from a developer installation.
