### 2018-08-03 #717 @jarvist
Mainly a question for the editor @kyleniemeyer, currently `paper.md` lists David Scanlon (the PI) as final + corresponding author. This would be standard for a scientific publication, but (from the commit history) they do not appear to have directly contributed to the code.
The current JOSS guidelines call for "A list of the authors of the software" (https://joss.readthedocs.io/en/latest/review_criteria.html#the-joss-paper), does this include non directly code-writing authors who may have contributed to the design of the package etc. ?

Generally I thought `paper.md` was useful + well written, but I think it could do with an extra sentence or two of introduction, similar to how `README.md` now reads. It just needs a bit more of an introduction explaining that these tools are for solid-state / band-structure calculations, making the 'statement of need' more clear.
