### 2018-08-03 #717 @jarvist
Re: tests
I'd recommend merging 'Developer installation' with 'Tests', and making it clear that a developer installation is a pre-requisite:
 https://github.com/SMTG-UCL/sumo/tree/6de16a00e2a488180bfafc34439d969737bd1854#developer-installation

The text around there could be improved - from inspecting the pip command, you are assuming that the reader would first clone the GitHub repository, and then run the pip command from that location as pwd.
Similar with the preceding sentence on building the documentation.
