### 2018-08-03 4c29d91
DOC: Improve README clarity re: dev installation
