### 2018-08-03 #717 @ajjackson
Good point! Have reworked the README: *Tests* and *Documentation* are now subsections of *Developer installation* and are explicit that you need to have the files.
