### 2018-08-03 #717 @kyleniemeyer
@jarvist regarding the authorship question, non-code intellectual contributions make someone eligible for authorship, as long as all authors agree on the author list.
