### 2018-08-03 a438202
DOC: Update paper
