### 2018-08-03 #717 @utf
@jarvist we've updated the paper to clarify in several places, including the paper title, that the the code is specifically for solid-state materials analysis.

Does that look ok now?
