### 2018-08-03 #717 @jarvist
Adding an explicit:
```
git clone https://github.com/SMTG-UCL/sumo.git
cd sumo
```
To the quoted shell commands for the developer installation would make it even clearer.

But I'm happy to sign off on this JOSS review as done!
