### 2018-08-04 #717 @kyleniemeyer
@utf only one comment on the paper before we accept: the reference for Togo 2013 is missing some info, like a DOI or URL. Once you've added that and confirmed the paper looks all good, then please archive your entire repo and provide the associated DOI here.
