### 2018-08-06 95f1d48
DOC: More on developer installation
