### 2018-08-06 #717 @ajjackson
Done! Have also added a note to Dev installation that *"Regular users can skip this section!"* and added explanation of `--user` flag.
