+++ paper.bib
@@ -67,7 +67,8 @@ publisher = "http://www.p4vasp.at/"}
 @misc{spglib,
   title={{spglib, a C library for finding and handling crystal symmetries}},
   author={Togo, A},
-  year={2013}
+  year={2013},
+  publisher = {https://atztogo.github.io/spglib}
 }
 
 @article{matplotlib,
