### 2018-08-06 #717 @ajjackson
Re: Spglib there doesn't seem to be a permanent identifier, but I have added the website https://atztogo.github.io/spglib/. It's available in the waybackmachine, at least!
