+++ setup.py
@@ -18,7 +18,7 @@ with open('README.rst', 'r') as file:
 
 setup(
     name='sumo',
-    version='1.0.9',
+    version='1.0.10',
     description=('Heavy weight plotting tools for ab initio '
                  'solid-state calculations'),
     url='https://github.com/smtg-ucl/sumo',
