### 2018-08-06 #717 @utf
@kyleniemeyer We've made the changes you asked for. The repository DOI is 10.5281/zenodo.1338124 (https://doi.org/10.5281/zenodo.1338124).
