### 2018-08-06 60058ab
DOC: Add JOSS paper to README
