### 2018-08-07 #717 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
